added countries

+++ src/Controller/CountriesController.php
@@ -48,11 +48,17 @@ class CountriesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function view($id = null) {
-		$country = $this->Countries->get($id, [
-			'contain' => ['Creators', 'Modifiers', 'Cities', 'Ports']
-		]);
-		$this->set('country', $country);
-		$this->set('_serialize', ['country']);
+		if ($this->request->params['_ext']) {
+			$country = $this->Countries->get($id);
+			$this->set('country', $country);
+			$this->set('_serialize', ['country']);
+		} else {
+			$country = $this->Countries->get($id, [
+				'contain' => ['Creators', 'Modifiers', 'Cities', 'Ports']
+			]);
+			$this->set('country', $country);
+			$this->set('_serialize', ['country']);
+		}
 	}
 
 /**
@@ -61,18 +67,33 @@ class CountriesController extends AppController {
  * @return void
  */
 	public function add() {
-		$country = $this->Countries->newEntity($this->request->data);
-		if ($this->request->is('post')) {
-			if ($this->Countries->save($country)) {
-				$this->Flash->success('The country has been saved.');
-				return $this->redirect(['action' => 'index']);
+		if ($this->request->params['_ext']) {
+			$country = $this->Countries->newEntity($this->request->data);
+			if ($this->Countries->save($country, ['validate' => false])) {
+				$message = 'Saved';
 			} else {
-				$this->Flash->error('The country could not be saved. Please, try again.');
+				$message = 'Error';
 			}
+			$this->set([
+				'data' => $this->request->data,
+				'message' => $message,
+				'country' => $country,
+				'_serialize' => ['message', 'country', 'data']
+			]);
+		} else {
+			$country = $this->Countries->newEntity($this->request->data);
+			if ($this->request->is('post')) {
+				if ($this->Countries->save($country)) {
+					$this->Flash->success('The country has been saved.');
+					return $this->redirect(['action' => 'index']);
+				} else {
+					$this->Flash->error('The country could not be saved. Please, try again.');
+				}
+			}
+			$creators = $this->Countries->Creators->find('list');
+			$modifiers = $this->Countries->Modifiers->find('list');
+			$this->set(compact('country', 'creators', 'modifiers'));
 		}
-		$creators = $this->Countries->Creators->find('list');
-		$modifiers = $this->Countries->Modifiers->find('list');
-		$this->set(compact('country', 'creators', 'modifiers'));
 	}
 
 /**
@@ -83,21 +104,40 @@ class CountriesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function edit($id = null) {
-		$country = $this->Countries->get($id, [
-			'contain' => []
-		]);
-		if ($this->request->is(['patch', 'post', 'put'])) {
-			$country = $this->Countries->patchEntity($country, $this->request->data);
-			if ($this->Countries->save($country)) {
-				$this->Flash->success('The country has been saved.');
-				return $this->redirect(['action' => 'index']);
-			} else {
-				$this->Flash->error('The country could not be saved. Please, try again.');
+		if ($this->request->params['_ext']) {
+			$country = $this->Countries->get($id);
+			if ($this->request->is(['patch', 'post', 'put'])) {
+				$country = $this->Countries->patchEntity($country, $this->request->data);
+				if ($this->Countries->save($country, ['validate' => false])) {
+					$message = 'Saved';
+				} else {
+					$message = 'Error';
+				}
 			}
+			$this->set([
+				'country' => $country,
+				'message' => $message,
+				'data' => $this->request->data,
+				'_serialize' => ['message','country','data']
+			]);
 		}
-		$creators = $this->Countries->Creators->find('list');
-		$modifiers = $this->Countries->Modifiers->find('list');
-		$this->set(compact('country', 'creators', 'modifiers'));
+		else {
+			$country = $this->Countries->get($id, [
+				'contain' => []
+			]);
+			if ($this->request->is(['patch', 'post', 'put'])) {
+				$country = $this->Countries->patchEntity($country, $this->request->data);
+				if ($this->Countries->save($country)) {
+					$this->Flash->success('The country has been saved.');
+					return $this->redirect(['action' => 'index']);
+				} else {
+					$this->Flash->error('The country could not be saved. Please, try again.');
+				}
+			}
+			$creators = $this->Countries->Creators->find('list');
+			$modifiers = $this->Countries->Modifiers->find('list');
+			$this->set(compact('country', 'creators', 'modifiers'));
+		}
 	}
 
 /**
@@ -108,13 +148,26 @@ class CountriesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function delete($id = null) {
-		$country = $this->Countries->get($id);
-		$this->request->allowMethod(['post', 'delete']);
-		if ($this->Countries->delete($country)) {
-			$this->Flash->success('The country has been deleted.');
-		} else {
-			$this->Flash->error('The country could not be deleted. Please, try again.');
+		if ($this->request->params['_ext']) {
+			$country = $this->Countries->get($id);
+			$message = 'Deleted';
+			if (!$this->Countries->delete($country)) {
+				$message = 'Error';
+			}
+			$this->set([
+				'message' => $message,
+				'_serialize' => ['message']
+			]);
 		}
-		return $this->redirect(['action' => 'index']);
+		else {		
+			$country = $this->Countries->get($id);
+			$this->request->allowMethod(['post', 'delete']);
+			if ($this->Countries->delete($country)) {
+				$this->Flash->success('The country has been deleted.');
+			} else {
+				$this->Flash->error('The country could not be deleted. Please, try again.');
+			}
+			return $this->redirect(['action' => 'index']);
+		}
 	}
 }
